Modularize the new state
