+++ client/state/editor/convert-to-blocks-dialog/actions.js
@@ -2,6 +2,7 @@
  * Internal dependencies
  */
 import { CONVERT_TO_BLOCKS_DIALOG_DISMISS } from 'state/action-types';
+import 'state/editor/init';
 
 export const dismissConvertToBlocksDialog = ( siteId, postId ) => ( {
 	type: CONVERT_TO_BLOCKS_DIALOG_DISMISS,
