+++ client/state/editor/convert-to-blocks-dialog/selectors.js
@@ -1,0 +1,13 @@
+/**
+ * External dependencies
+ */
+import { get } from 'lodash';
+
+/**
+ * Internal dependencies
+ */
+import 'state/editor/init';
+
+export function isConvertToBlocksDialogDismissed( state, siteId, postId ) {
+	return 'dismissed' === get( state, [ 'editor', 'convertToBlocksDialog', siteId, postId ] );
+}
