--- client/state/selectors/is-convert-to-blocks-dialog-dismissed.js
@@ -1,8 +1,0 @@
-/**
- * External dependencies
- */
-import { get } from 'lodash';
-
-export default function isConvertToBlocksDialogDismissed( state, siteId, postId ) {
-	return 'dismissed' === get( state, [ 'ui', 'convertToBlocksDialog', siteId, postId ] );
-}
