+++ client/state/ui/reducer.js
@@ -11,7 +11,6 @@ import { combineReducers, withoutPersistence } from 'state/utils';
 import actionLog from './action-log/reducer';
 import checkout from './checkout/reducer';
 import comments from './comments/reducer';
-import convertToBlocksDialog from './convert-to-blocks-dialog/reducer';
 import dropZone from './drop-zone/reducer';
 import guidedTour from './guided-tours/reducer';
 import editorDeprecationDialog from './editor-deprecation-dialog/reducer';
@@ -97,7 +96,6 @@ const reducer = combineReducers( {
 	actionLog,
 	checkout,
 	comments,
-	convertToBlocksDialog,
 	dropZone,
 	editorDeprecationDialog,
 	guidedTour,
